requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.14.1=pypi_0
aiohttp=3.7.4.post0=pypi_0
alembic=1.4.1=pypi_0
argon2-cffi=20.1.0=py39h27cfd23_1
asdfghjkl=0.1a2=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=3.0.1=pypi_0
async_generator=1.10=py_0
attr=0.3.1=pypi_0
attrs=21.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backpack-for-pytorch=1.5.0=pypi_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=4.1.0=pyhd8ed1ab_0
bottleneck=1.3.2=py39hdd57654_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py39hf3d152e_1
cffi=1.14.6=py39h400218f_0
chardet=4.0.0=py39h06a4308_1003
click=8.0.3=pypi_0
cloudpickle=2.0.0=pypi_0
colorama=0.4.4=pypi_0
conda=4.11.0=py39hf3d152e_0
conda-package-handling=1.7.3=py39h27cfd23_1
configparser=5.0.2=pypi_0
cryptography=3.4.7=py39hd23ed53_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.10.0=pypi_0
databricks-cli=0.15.0=pypi_0
dataclasses=0.6=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.0=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
deprecated=1.2.13=pypi_0
descartes=1.1.0=pypi_0
dm-tree=0.1.6=pypi_0
docker=5.0.3=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
einops=0.4.0=pypi_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
filelock=3.4.0=pypi_0
flask=2.0.2=pypi_0
flatbuffers=2.0=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
fsspec=2021.10.1=pypi_0
future=0.18.2=pypi_0
gast=0.4.0=pypi_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.7=pypi_0
gitpython=3.1.24=pypi_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.3.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googledrivedownloader=0.4=pyhd3deb0d_1
gpflow=2.3.0=pypi_0
greenlet=1.1.2=pypi_0
grpcio=1.41.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gunicorn=20.1.0=pypi_0
gust=0.1=pypi_0
h5py=3.6.0=pypi_0
huggingface-hub=0.2.1=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=4.10.0=py39hf3d152e_0
importlib_resources=5.4.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=5.5.5=py39hef51801_0
ipython=7.31.0=py39hf3d152e_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.5=pypi_0
itsdangerous=2.0.1=pypi_0
jedi=0.18.1=py39hf3d152e_0
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonpickle=1.5.2=pypi_0
jsonschema=4.3.3=pyhd8ed1ab_0
jupyter_client=7.1.0=pyhd8ed1ab_0
jupyter_core=4.9.1=py39hf3d152e_1
jupyterlab-widgets=1.0.2=pypi_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
keras=2.7.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
laplace-torch=0.1a2=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libclang=12.0.0=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
littleutils=0.2.2=pypi_0
llvmlite=0.37.0=pypi_0
lz4-c=1.9.3=h295c915_1
mako=1.1.5=pypi_0
markdown=3.3.4=pypi_0
markupsafe=2.0.1=py39h3811e60_0
matplotlib=3.4.3=pypi_0
matplotlib-base=3.5.0=py39h3ed280b_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py39h3811e60_1004
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mlflow=1.20.2=pypi_0
multidict=5.2.0=pypi_0
multipledispatch=0.6.0=pypi_0
munch=2.5.0=pypi_0
munkres=1.1.4=py_0
nbclient=0.5.9=pyhd8ed1ab_0
nbconvert=6.4.0=py39hf3d152e_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.4=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pyhd8ed1ab_1
nflows=0.14=pypi_0
ninja=1.10.2.2=pypi_0
nltk=3.6.5=pypi_0
notebook=6.4.6=pyha770c72_0
numba=0.54.1=pypi_0
numexpr=2.8.1=py39h6abb31d_0
numpy=1.20.3=pypi_0
oauthlib=3.1.1=pypi_0
ogb=1.3.2=pypi_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1l=h7f8727e_0
opentsne=0.6.1=pypi_0
opt-einsum=3.3.0=pypi_0
outdated=0.2.1=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=pypi_0
pandoc=2.16.2=h7f98852_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pathtools=0.1.2=pypi_0
pcre=8.45=h295c915_0
pdf2image=1.16.0=pypi_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.4.0=py39h5aabda8_0
pip=21.1.3=py39h06a4308_0
plotly=5.5.0=pypi_0
prometheus-flask-exporter=0.18.3=pypi_0
prometheus_client=0.12.0=pyhd8ed1ab_0
promise=2.3=pypi_0
prompt-toolkit=3.0.24=pyha770c72_0
protobuf=3.18.1=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
py-cpuinfo=8.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyblaze=2.3.2=pypi_0
pycosat=0.6.3=py39h27cfd23_0
pycparser=2.20=py_2
pydeprecate=0.3.1=pypi_0
pyg=2.0.3=py39_torch_1.10.0_cu113
pygments=2.11.2=pyhd8ed1ab_0
pymongo=4.0.1=pypi_0
pynndescent=0.5.5=pypi_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=3.0.6=pypi_0
pyqt=5.9.2=py39h2531618_6
pyro-api=0.1.2=pypi_0
pyro-ppl=1.8.0=pypi_0
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.5=h12debd9_4
python-dateutil=2.8.2=pyhd3eb1b0_0
python-editor=1.0.4=pypi_0
python-louvain=0.15=pyhd3deb0d_0
python_abi=3.9=2_cp39
pytorch=1.10.1=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-cluster=1.5.9=py39_torch_1.10.0_cu113
pytorch-lightning=1.5.10=pypi_0
pytorch-mutex=1.0=cuda
pytorch-scatter=2.0.9=py39_torch_1.10.0_cu113
pytorch-sparse=0.6.12=py39_torch_1.10.0_cu113
pytorch-spline-conv=1.2.1=py39_torch_1.10.0_cu113
pytz=2021.3=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py39hb69f2a1_2
qt=5.9.7=h5867ecd_1
querystring-parser=1.2.4=pypi_0
readline=8.1=h27cfd23_0
regex=2021.11.10=pypi_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
ruamel_yaml=0.15.100=py39h27cfd23_0
sacred=0.8.2=pypi_0
sacremoses=0.0.47=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.7.3=py39hc147768_0
seaborn=0.11.2=pyhd3eb1b0_0
seml=0.3.6=dev_0
send2trash=1.8.0=pyhd8ed1ab_0
sentence-transformers=2.1.0=pypi_0
sentencepiece=0.1.96=pypi_0
sentry-sdk=1.4.3=pypi_0
setproctitle=1.2.2=pypi_0
setuptools=59.5.0=pypi_0
shapely=1.8.1.post1=pypi_0
shortuuid=1.0.1=pypi_0
sip=4.19.13=py39h2531618_0
six=1.16.0=pyhd3eb1b0_0
smmap=4.0.0=pypi_0
sqlalchemy=1.4.25=pypi_0
sqlite=3.36.0=hc218d9a_0
sqlparse=0.4.2=pypi_0
subprocess32=3.5.4=pypi_0
tabulate=0.8.9=pypi_0
tenacity=8.0.1=pypi_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.7.0=pypi_0
tensorflow-estimator=2.7.0=pypi_0
tensorflow-io-gcs-filesystem=0.23.1=pypi_0
tensorflow-probability=0.15.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.12.1=py39hf3d152e_1
testpath=0.5.0=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tinydb=4.5.2=pypi_0
tinydb-serialization=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.3=pypi_0
torch-householder=1.0.1=pypi_0
torchaudio=0.10.1=py39_cu113
torchmetrics=0.5.1=pypi_0
torchvision=0.11.2=py39_cu113
tornado=6.1=py39h27cfd23_0
tqdm=4.61.2=pyhd3eb1b0_1
traitlets=5.1.1=pyhd8ed1ab_0
transformers=4.15.0=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021a=h52ac0ba_0
umap=0.1.1=pypi_0
umap-learn=0.5.2=pypi_0
urllib3=1.26.6=pyhd3eb1b0_1
wandb=0.12.11=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=1.2.1=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.2=pypi_0
wrapt=1.13.2=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.6=py_0
yaml=0.2.5=h7b6447c_0
yarl=1.7.0=pypi_0
yaspin=2.1.0=pypi_0
zeromq=4.3.4=h9c3ff4c_0
zipp=3.7.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
